Please provide the file changes or a description of the code modifications so I can generate the commit message for you.

+++ step17-board-backend/src/main/java/com/spring/mapper/BoardMapper.java
@@ -20,8 +20,8 @@ public interface BoardMapper {
 
   int boardCount();
 
-  BoardDTO selectBoard(Long bno);
+  BoardDTO selectBoard(int bno);
 
-  List<BoardCommentDTO> selectBoardComment(Long bno);
+  List<BoardCommentDTO> selectBoardComment(int bno);
 
 }
